# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Course API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api/"
        }
    ],
    "paths": {
        "/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get a list of all courses",
                "description": "Returns a list of all courses",
                "operationId": "getCoursesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Store a new course",
                "description": "Returns course data",
                "operationId": "storeCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get a single course by ID.",
                "description": "Returns course data",
                "operationId": "getCourseById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Update existing course",
                "description": "Returns updated course data",
                "operationId": "updateCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Delete existing course",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CourseRequest": {
                "title": "Store Course request",
                "description": "Store Course request body data",
                "required": [
                    "title",
                    "description"
                ],
                "properties": {
                    "title": {
                        "title": "Title",
                        "description": "Title of the new course",
                        "type": "string",
                        "example": "My Course"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the new course",
                        "type": "string",
                        "example": "This is new course's description"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-03-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-03-27 17:50:45"
                    },
                    "status": {
                        "title": "Course status.",
                        "description": "Course status.",
                        "type": "string",
                        "enum": [
                            "Pending",
                            "Published"
                        ]
                    },
                    "is_premium": {
                        "title": "Course premium flag",
                        "description": "Course premium flag",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Course": {
                "title": "Course",
                "description": "Course model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title of the new course",
                        "type": "string",
                        "example": "My Course"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the new course",
                        "type": "string",
                        "example": "This is new course's description"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-27 17:50:45"
                    },
                    "status": {
                        "title": "Course status.",
                        "description": "Course status.",
                        "type": "string",
                        "default": "Pending",
                        "enum": [
                            "Pending",
                            "Published"
                        ]
                    },
                    "is_premium": {
                        "title": "Course premium flag",
                        "description": "Course premium flag",
                        "type": "boolean",
                        "format": "int64",
                        "default": false
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Course"
                }
            },
            "CourseResource": {
                "title": "CourseResource",
                "description": "Course resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Course"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CourseResource"
                }
            }
        }
    }
}